refactor: 요일 enum으로 정리
