--- src/main/java/attendance/domain/Week.java
@@ -1,45 +1,0 @@
-package attendance.domain;
-
-import java.time.DayOfWeek;
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-public enum Week {
-    MONDAY(DayOfWeek.MONDAY, LocalTime.of(13, 0)),
-    TUESDAY(DayOfWeek.TUESDAY, LocalTime.of(10, 0)),
-    WEDNESDAY(DayOfWeek.WEDNESDAY, LocalTime.of(10, 0)),
-    THURSDAY(DayOfWeek.THURSDAY, LocalTime.of(10, 0)),
-    FRIDAY(DayOfWeek.FRIDAY, LocalTime.of(10, 0));
-
-    private final DayOfWeek dayOfWeek;
-    private final LocalTime startTime;
-
-    Week(DayOfWeek dayOfWeek, LocalTime startTime) {
-        this.dayOfWeek = dayOfWeek;
-        this.startTime = startTime;
-    }
-
-    public DayOfWeek getDayOfWeek() {
-        return dayOfWeek;
-    }
-
-    public LocalTime getStartTime() {
-        return startTime;
-    }
-
-    public static boolean isAttendanceDay(LocalDate date) {
-        DayOfWeek dayOfWeek = date.getDayOfWeek();
-        return dayOfWeek != DayOfWeek.SATURDAY &&
-                dayOfWeek != DayOfWeek.SUNDAY;
-    }
-
-    public static Week from(LocalDate date) {
-        DayOfWeek dayOfWeek = date.getDayOfWeek();
-        for (Week day : values()) {
-            if (day.dayOfWeek == dayOfWeek) {
-                return day;
-            }
-        }
-        throw new IllegalArgumentException("출석일이 아닙니다: " + date);
-    }
-}
