+++ src/main/java/attendance/domain/model/AttendanceDate.java
@@ -1,0 +1,28 @@
+package attendance.domain.model;
+
+import java.time.LocalDate;
+import attendance.domain.enums.*;
+
+public class AttendanceDate {
+    private final LocalDate value;
+
+    public AttendanceDate(LocalDate value) {
+        this.value = value;
+    }
+
+    public AttendanceDate(int day) {
+        this.value = LocalDate.of(2024, 12, day);
+    }
+
+    public LocalDate getValue() {
+        return value;
+    }
+
+    public String getDayOfWeekName() {
+        return Week.getKoreanName(value.getDayOfWeek());
+    }
+
+    public boolean isAttendanceDay() {
+        return Week.isAttendanceDay(value);
+    }
+}
